+++ src/main/java/dao/ConnectionDao.java
@@ -1,0 +1,49 @@
+package dao;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class ConnectionDao {
+	private static final ConnectionDao CONNECTION_DAO;
+
+	static {
+		CONNECTION_DAO = new ConnectionDao();
+	}
+
+	private ConnectionDao() {
+	}
+
+	public static ConnectionDao getInstance() {
+		return CONNECTION_DAO;
+	}
+
+	public Connection getConnection() {
+		Connection connection = null;
+		final String server = "localhost:13306"; // MySQL 서버 주소
+		final String database = "woowachess"; // MySQL DATABASE 이름
+		final String option = "?useSSL=false&serverTimezone=UTC&characterEncoding=utf8";
+		final String userName = "root"; //  MySQL 서버 아이디
+		final String password = "root"; // MySQL 서버 비밀번호
+
+		// 드라이버 로딩
+		try {
+			Class.forName("com.mysql.cj.jdbc.Driver");
+		} catch (ClassNotFoundException e) {
+			System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
+			e.printStackTrace();
+		}
+
+		// 드라이버 연결
+		try {
+			connection = DriverManager.getConnection(
+					"jdbc:mysql://" + server + "/" + database + option, userName, password);
+			System.out.println("정상적으로 연결되었습니다.");
+		} catch (SQLException e) {
+			System.err.println("연결 오류:" + e.getMessage());
+			e.printStackTrace();
+		}
+
+		return connection;
+	}
+}
